+++ pkg/function/server/server_test.go
@@ -9,7 +9,6 @@ import (
 	"time"
 
 	functionpb "github.com/numaproj/numaflow-go/pkg/apis/proto/function/v1"
-	"github.com/numaproj/numaflow-go/pkg/configs"
 	functionsdk "github.com/numaproj/numaflow-go/pkg/function"
 	"github.com/numaproj/numaflow-go/pkg/function/client"
 	"github.com/stretchr/testify/assert"
@@ -59,7 +58,7 @@ func Test_server_map(t *testing.T) {
 			for i := 0; i < 10; i++ {
 				key := fmt.Sprintf("client_%d", i)
 				// set the key in metadata for map function
-				md := metadata.New(map[string]string{configs.DatumKey: key})
+				md := metadata.New(map[string]string{functionsdk.DatumKey: key})
 				ctx = metadata.NewOutgoingContext(ctx, md)
 				list, err := c.MapFn(ctx, &functionpb.Datum{
 					Key:       key,
@@ -156,7 +155,7 @@ func Test_server_reduce(t *testing.T) {
 			close(reduceDatumCh)
 
 			// set the key in gPRC metadata for reduce function
-			md := metadata.New(map[string]string{configs.DatumKey: testKey})
+			md := metadata.New(map[string]string{functionsdk.DatumKey: testKey})
 			ctx = metadata.NewOutgoingContext(ctx, md)
 			list, err := c.ReduceFn(ctx, reduceDatumCh)
 			assert.NoError(t, err)
